% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Softball Webscraper}
\usage{
scrape(url)
}
\arguments{
\item{url}{the web address of the teams' data to scrape}
}
\value{
A dataframe consisting of each player's Number, NAME, AVG, 
OPS, AB, R, H, 2B, 3B, HR, RBI, TB, SLG% BB, HBP, SO, GDP, OB% SF, SH,
SB, ATT, GP, and GS.
}
\description{
Scrapes the player statistics from a given URL.
}
\examples{
url <-"https://wmubroncos.com/sports/softball/stats/2019"
test <- scrape(url)
test_probs <- prob_calc(test)
        
}
